requirements.txt
Flask==3.0.0
requests==2.31.0
markdown==3.5.1
python-dotenv==1.0.0
openai==1.14.0
anthropic==0.18.0
flask-session==0.5.0
reportlab==4.1.0